Remove legacy retry stages and supporting classes
