+++ core/sdk-core/src/main/java/software/amazon/awssdk/core/internal/http/pipeline/stages/RetryableStage.java
@@ -17,9 +17,9 @@ package software.amazon.awssdk.core.internal.http.pipeline.stages;
 
 import java.io.IOException;
 import java.time.Duration;
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import software.amazon.awssdk.annotations.SdkInternalApi;
-import software.amazon.awssdk.annotations.SdkTestInternalApi;
 import software.amazon.awssdk.core.Response;
 import software.amazon.awssdk.core.exception.SdkException;
 import software.amazon.awssdk.core.internal.http.HttpClientDependencies;
@@ -27,77 +27,119 @@ import software.amazon.awssdk.core.internal.http.RequestExecutionContext;
 import software.amazon.awssdk.core.internal.http.pipeline.RequestPipeline;
 import software.amazon.awssdk.core.internal.http.pipeline.RequestToResponsePipeline;
 import software.amazon.awssdk.core.internal.http.pipeline.stages.utils.RetryableStageHelper;
-import software.amazon.awssdk.core.internal.retry.RateLimitingTokenBucket;
 import software.amazon.awssdk.http.SdkHttpFullRequest;
+import software.amazon.awssdk.http.SdkHttpFullResponse;
 
 /**
  * Wrapper around the pipeline for a single request to provide retry, clock-skew and request throttling functionality.
  */
 @SdkInternalApi
 public final class RetryableStage<OutputT> implements RequestToResponsePipeline<OutputT> {
+    private static final String RETRY_AFTER_HEADER = "Retry-After";
     private final RequestPipeline<SdkHttpFullRequest, Response<OutputT>> requestPipeline;
     private final HttpClientDependencies dependencies;
-    private final RateLimitingTokenBucket rateLimitingTokenBucket;
 
     public RetryableStage(HttpClientDependencies dependencies,
                           RequestPipeline<SdkHttpFullRequest, Response<OutputT>> requestPipeline) {
         this.dependencies = dependencies;
         this.requestPipeline = requestPipeline;
-        this.rateLimitingTokenBucket = null;
-    }
-
-    @SdkTestInternalApi
-    public RetryableStage(HttpClientDependencies dependencies,
-                          RequestPipeline<SdkHttpFullRequest, Response<OutputT>> requestPipeline,
-                          RateLimitingTokenBucket rateLimitingTokenBucket) {
-        this.dependencies = dependencies;
-        this.requestPipeline = requestPipeline;
-        this.rateLimitingTokenBucket = rateLimitingTokenBucket;
     }
 
     @Override
     public Response<OutputT> execute(SdkHttpFullRequest request, RequestExecutionContext context) throws Exception {
-        RetryableStageHelper retryableStageHelper = new RetryableStageHelper(request, context, rateLimitingTokenBucket,
-                                                                             dependencies);
-
+        RetryableStageHelper retryableStageHelper = new RetryableStageHelper(request, context, dependencies);
+        Duration initialDelay = retryableStageHelper.acquireInitialToken();
+        TimeUnit.MILLISECONDS.sleep(initialDelay.toMillis());
         while (true) {
-            retryableStageHelper.startingAttempt();
-
-            if (!retryableStageHelper.retryPolicyAllowsRetry()) {
-                throw retryableStageHelper.retryPolicyDisallowedRetryException();
-            }
-
-            retryableStageHelper.getSendToken();
-
-            Duration backoffDelay = retryableStageHelper.getBackoffDelay();
-            if (!backoffDelay.isZero()) {
-                retryableStageHelper.logBackingOff(backoffDelay);
-                TimeUnit.MILLISECONDS.sleep(backoffDelay.toMillis());
-            }
-
-            Response<OutputT> response;
             try {
-                retryableStageHelper.logSendingRequest();
-                response = requestPipeline.execute(retryableStageHelper.requestToSend(), context);
-            } catch (SdkException | IOException e) {
-                retryableStageHelper.setLastException(e);
-                retryableStageHelper.updateClientSendingRateForErrorResponse();
-                continue;
+                retryableStageHelper.startingAttempt();
+                Response<OutputT> response = executeRequest(retryableStageHelper, context);
+                retryableStageHelper.recordAttemptSucceeded();
+                return response;
+            } catch (SdkExceptionWithRetryAfterHint | SdkException | IOException e) {
+                Throwable throwable = e;
+                if (e instanceof SdkExceptionWithRetryAfterHint) {
+                    SdkExceptionWithRetryAfterHint wrapper = (SdkExceptionWithRetryAfterHint) e;
+                    throwable = wrapper.cause();
+                }
+                retryableStageHelper.setLastException(throwable);
+                Duration suggestedDelay = suggestedDelay(e);
+                Optional<Duration> backoffDelay = retryableStageHelper.tryRefreshToken(suggestedDelay);
+                if (backoffDelay.isPresent()) {
+                    Duration delay = backoffDelay.get();
+                    retryableStageHelper.logBackingOff(delay);
+                    TimeUnit.MILLISECONDS.sleep(delay.toMillis());
+                } else {
+                    throw retryableStageHelper.retryPolicyDisallowedRetryException();
+                }
             }
+        }
+    }
 
-            retryableStageHelper.setLastResponse(response.httpResponse());
+    private Duration suggestedDelay(Exception e) {
+        if (e instanceof SdkExceptionWithRetryAfterHint) {
+            SdkExceptionWithRetryAfterHint except = (SdkExceptionWithRetryAfterHint) e;
+            return Duration.ofSeconds(except.retryAfter());
+        }
+        return Duration.ZERO;
+    }
 
-            if (!response.isSuccess()) {
-                retryableStageHelper.adjustClockIfClockSkew(response);
-                retryableStageHelper.setLastException(response.exception());
-                retryableStageHelper.updateClientSendingRateForErrorResponse();
-                continue;
+    /**
+     * Executes the requests and returns the result. If the response is not successful throws the wrapped exception.
+     */
+    private Response<OutputT> executeRequest(RetryableStageHelper retryableStageHelper,
+                                             RequestExecutionContext context) throws Exception {
+        retryableStageHelper.logSendingRequest();
+        Response<OutputT> response = requestPipeline.execute(retryableStageHelper.requestToSend(), context);
+        retryableStageHelper.setLastResponse(response.httpResponse());
+        if (!response.isSuccess()) {
+            retryableStageHelper.adjustClockIfClockSkew(response);
+            throw responseException(response);
+        }
+        return response;
+    }
+
+    private RuntimeException responseException(Response<OutputT> response) {
+        Optional<Integer> optionalRetryAfter = retryAfter(response.httpResponse());
+        if (optionalRetryAfter.isPresent()) {
+            return new SdkExceptionWithRetryAfterHint(optionalRetryAfter.get(), response.exception());
+        }
+        return response.exception();
+    }
+
+    private Optional<Integer> retryAfter(SdkHttpFullResponse response) {
+        Optional<String> optionalRetryAfterHeader = response.firstMatchingHeader(RETRY_AFTER_HEADER);
+        if (optionalRetryAfterHeader.isPresent()) {
+            String retryAfterHeader = optionalRetryAfterHeader.get();
+            try {
+                return Optional.of(Integer.parseInt(retryAfterHeader));
+            } catch (NumberFormatException e) {
+                // Ignore and fallback to returning empty.
             }
+        }
+        return Optional.empty();
+    }
 
-            retryableStageHelper.updateClientSendingRateForSuccessResponse();
+    // This probably should go directly into SdkException
+    static class SdkExceptionWithRetryAfterHint extends RuntimeException {
+        private final SdkException cause;
+        private final int seconds;
 
-            retryableStageHelper.attemptSucceeded();
-            return response;
+        SdkExceptionWithRetryAfterHint(int seconds, SdkException cause) {
+            this.seconds = seconds;
+            this.cause = cause;
+        }
+
+        public int retryAfter() {
+            return seconds;
+        }
+
+        public SdkException cause() {
+            return cause;
+        }
+
+        public int seconds() {
+            return seconds;
         }
     }
 }
